refactor(mito2): use absolute scheduler imports
